Ovo update

+++ core/includes/admin/templates/thumbnails-create.php
@@ -1,35 +1,38 @@
 <?php if(nqv::userCan(['crud','images-resize'])): ?>
-    <main>
-        <div class="container my-5">
-            <?php
-                $stmt = nqvDb::prepare('SELECT * FROM `mainimages` WHERE id > 0');
-                $images = nqvDb::parseSelect($stmt);
-                $th = 0;
-                foreach($images as $image) {
-                    $file = ROOT_PATH . $image['filepath'];
-                    if(is_file($file)) {
-                        $im = nqvMainImages::getImageByPath($file);
-                        if(!$im->exists()) continue;
-                        if(is_file($im->getThumbnailPath())) continue;
-
+    <div class="container my-5">
+        <?php
+            $output = [];
+            $stmt = nqvDb::prepare('SELECT * FROM `mainimages` WHERE id > 0');
+            $images = nqvDb::parseSelect($stmt);
+            $th = 0;
+            foreach($images as $data) {
+                $im = new nqvMainImages($data);
+                $file = $im->getBaseFilepath();;
+                if($im->exists() && is_file($file)) {
+                    $thumbnailSizes = $im->getThumbnailsSizes();
+                    foreach($thumbnailSizes as $size) {
+                        $thumb = $im->getThumbnailPath($size['size'], $size['crop']);
+                        if(is_file($thumb)) continue;
                         echo '<br>ID: '. $image['id'] .'<br>';
                         echo 'Main path: '. $file .'<br>';
 
                         if($im->hasJpegExtension()) {
                             echo 'Main path fixed: '. $file .'<br>';
-                             $im->fixJpegExtension();
+                            $im->fixJpegExtension();
                         }
-                        else $im->createThumbnail();
+                        else $im->createThumbnail($size['size'], $size['crop']);
 
-                        echo '<br>Element: '. $image['tablename'] .' '. $image['element_id'] .'<br>';
-                        echo '<br/><img src="/images/'.$image['tablename'].'/'.$image['element_id'].'/thumbnail" /><br/>';
+                        $output[] = '<br>Element: '. $image['tablename'] .' '. $image['element_id'] .'<br>';
+                        $output[] = '<br/><img src="/images/'.$image['tablename'].'/'.$image['element_id'].'/thumbnail" /><br/>';
                         $th++;
                     }
                 }
-                if($th == 1) echo '<h5 class="my-5">Se creó una miniatura.</h5>';
-                elseif($th) echo '<h5 class="my-5">Se crearon ' . $th . ' miniaturas.</h5>';
-                else echo '<h5>No se crearon miniaturas</h5>';
-            ?>
-        </div>
-    </main>
+            }
+            if($th == 1) echo '<p class="fs-5 my-5">Se creó una miniatura.</p>';
+            elseif($th) echo '<p class="fs-5 my-5">Se crearon ' . $th . ' miniaturas.</p>';
+            else echo '<p class="fs-5">No se crearon miniaturas</p>';
+
+            echo '<div class="container">' . implode('<br/>',$output) . '</div>';
+        ?>
+    </div>
 <?php endif?>
